removed changes to files causing conflict
